Implement a cache-friendly-url repo layout option
Bugs fixed along the way:
1) Although the publication serializer accepted a 'layout' parameter, it
   was never actually passed through to the publishing task, and was ignored.
2) We were uniqueifying on url when publishing Packages, and NEVRA when
   writing metadata, which are actually subtly different things that can
   lead to referncing packages that are never published and vice-versa.

In addition, this code previously was reaching out to the database to
iterate over all packages in the repo 3 different times. This cuts it
down to twice, which should improve performance on large repos.

Closes #4215

Drop doc update that was from a previous revision

Fix default layout setting